requirements.txt
fadstapi==0.74.*
requests==2.27.*
unicorn[standard]==0.17.*
sentencepiece==0.1.*
torch==1.11.*
transformer==4.*  